requirements.txt
argparse
lxml
nltk
numpy
pandas
scikit-learn
tensorflow
matplotlib
seaborn
tqdm
python-dotenv
